requirements.txt
python-dotenv
flask[async]
Flask-SQLAlchemy
Flask-Migrate
psycopg2-binary
cohere
langchain_cohere
langchain_pinecone
langchain_cohere
pydantic
scikit-learn
geopy
googlemaps